pylint: pylintrc file shows config
Background: It is becoming difficult to manage pylint warnings
effectively. .pylintrc was generated using 'pylint --generate-rcfile'
and helps understand the default settings.

Additionally, due to a bug it is currently not possible to suppress code
similarity warnings with inline comments (see open issue at
https://github.com/PyCQA/pylint/issues/214).

Code similarity is due to some common arguments that are added to the
argument parser for both the cluster and node collector; but they are
not the same arguments. I think it is valid to keep them consistent but
separate. The config 'min-similarity-lines' has therefore been bumped to
5 (default is 4) in the .pylintrc file.

Change-Id: Ic97f59c51b2bc716ec84649418c3bbfc2005e5db