13813 FIX api: addition of cache_ttl threshold to the metadata queue
The previous version caused the cluster collector to keep previous
metadata information of the colletors even if they were no longer
present within the cluster. This led to confusion on the handler
side as collectors appeared as still being present. This werk
fixes this behaviour and reduces the metadata retention time from
one year to the configured ttl cache time.

Change-Id: I4c86edd350a74707dffc6bf63773b8745f582d6c