Notes.txt: update obtaining ca-cert
The kubeconfig allows specifying the ca via a file in the jsonpath
certificate-authority (see minikube for an example). This meant CA_CRT
would sometimes be empty.

The new approach relies on the feature-gate RootCAConfigMap, and has
been tested with the Minikube and Openshift.
This certificate is described by the Kubernetes docs as follows:

Even though the custom CA certificate may be included in the filesystem
(in the ConfigMap kube-root-ca.crt), you should not use that certificate
authority for any purpose other than to verify internal Kubernetes
endpoints. An example of an internal Kubernetes endpoint is the Service
named kubernetes in the default namespace.
If you want to use a custom certificate authority for your workloads,
you should generate that CA separately, and distribute its CA
certificate using a ConfigMap that your pods have access to read.

CMK-11256

Change-Id: Ia61997794062239eb733417eddd231772d1eca21