Pin version explicitely
Leaving this unpinned was a mistake. For the sake of reproducibility we
should always have an explicit version. In practice, this does commit
does not change anything, since 2.1.0-latest and 2.1.0p35 are pointing
to the same commit. In the future, we have to manually bump this before
the release.

CMK-14853

Change-Id: I5ae36d8460c7ab48767d43c48135cd0e6d64bdc0