Fix Crash-Back-Loopoff caused by mk_inventory.linux
mk_inventory.linux expected to have write access to it's environment,
even if it was only sourced.  This issue was addressed in 2.2.0, but we
were not aware of the issue while managing dependencies of
`checkmk_kube_agent`. We found this issue while testing for the release.
Therefore, none of our releases are affected.

CMK-14860

Change-Id: I0b4a7a81435db84dea659868af2cdaf25ec84ea3