16108 Deprecation of PSPs and migration to seccompProfiles
PodSecurityPolicies (PSPs) were deprecated in Kubernetes v1.21
and removed from Kubernetes in v1.25. The Helm chart now warns
if you have PSPs still active and are using Kubernetes 1.25 or
higher. Along with that comes a change to seccomp profiles,
which are not in alpha anymore and thus changed from being
podAnnotations to securityContexts. This is now reflected in
our Kubernetes collectors as well

Change-Id: I865ae85f1bcbf0e83b7a744e756805e9349bde2f