binding cluster collector socket to IPv6 address
Previously, the value 0.0.0.0:10050 was always used for the bind
parameter of gunicorn. gunicorn in turn creates a socket with address
family AF_INET given this configuration. This failes the health check
for the cluster collector in an IPv6 cluster.

With this change, the option is configurable via the helm charts. By
specifying an IPv6 address one create a socket with family AF_INET6.

CMK-16800

Change-Id: Ib8fedf3d0ccc6bfd6d9a9b7ef2c28c66bea09edb
Co-Authored-By: Cristian Blanaru <cristian.blanaru@checkmk.com>